Removed some threads for a single loop
The threads that handle the arm and the use inactive watchdog have been removed. The main thread now handles them in a loop with the kiosk system.

To allow the loop running all the time the command queue has been modified. Now it's not blocking anymore. Being called at every cycle, the queue returns either a command or nothing.

+++ src/main/core/ports/IArmPort.h
@@ -5,12 +5,11 @@
 #ifndef GRABSTATION_ARMPORT_H
 #define GRABSTATION_ARMPORT_H
 #pragma once
-#include <atomic>
 
 struct Coordinate
 {
-    int x;
-    int y;
+    float x;
+    float y;
 };
 
 /**
